+++ app/tests/test_views_templates.py
@@ -40,6 +40,14 @@ def describe_detail():
                 "_self": "http://localhost:5000/templates/iw",
             }
 
+        def it_defaults_to_gif_example_when_available(expect, client):
+            request, response = client.get("/templates/bongo")
+            expect(response.status) == 200
+            expect(response.json["example"]["url"]) == (
+                "http://localhost:5000/images/bongo/"
+                "Any_sound_when_you're_trying_to_sleep/Max_volume_alarm_when_you_have_to_wake_up.gif"
+            )
+
         def it_returns_404_when_missing(expect, client):
             request, response = client.get("/templates/foobar")
             expect(response.status) == 404
@@ -71,7 +79,7 @@ def describe_detail():
                 "?background=https://www.gstatic.com/webp/gallery/3.png"
             }
 
-        @pytest.mark.parametrize("id", ["fry", "custom"])
+        @pytest.mark.parametrize("id", ["iw", "custom"])
         def it_redirects_if_requested(expect, client, id):
             data = {"text_lines": ["abc"], "redirect": True}
             request, response = client.post(